requirements.txt
asgiref==3.2.10
autopep8==1.7.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-oauth==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
Pillow==9.1.0
pip-search==0.0.12
pycodestyle==2.9.1
Pygments==2.13.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
rich==12.5.1
simplejson==3.17.6
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
